Import/export of irreductible representation associated to YoungTableau (#4)
The irreductible representation (U and V Csr objects) are build at runtime and stored in .bin file. The file is then loaded during compilation and used to fill a static constexpr YoungTableau::s_irrep object. Because this is constexpr, compiler is expected to perform optimizations in tensor products etc... that are not implemented yet.

- Rely on the #embed directive from C23 -> Only tested with clang 19
- Previous Csr class is renamed CsrDynamic because based on std::vectors. A new Csr class is based on std::array
- An irrep tag (string like 1_2l3d4 is generated at compile-time to identify the Young tableau.